feat: change axios config path, redux state name, navigate user after succeed register
