+++ src/components/firstViewContainer/tasksContainer/tasksContainer.scss
@@ -7,7 +7,7 @@
 
   .tasksContainer {
     margin: auto;
-    margin-top: 20px;
+    margin-top: 25px;
     min-height: 500px;
     min-width: 300px;
     max-width: 300px;
