+++ src/components/shared/logout/logoutContainer.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { useMutation } from 'react-query'
-import requestTaskApi from '../../config/axiosConfig'
+import requestTaskApi from '../../../config/axiosConfig'
 import './logoutContainer.scss'
 import { toast } from 'react-toastify'
 import { store } from '../../../redux/store'
-import { getClientResponse } from '../../../redux/clientResponse'
+import { getClientResponse } from '../../../redux/user'
 
 const LogoutContainer = () => {
 	const { mutate: logout } = useMutation(
