+++ src/components/subPages/loginForm/useLoginForm.ts
@@ -1,10 +1,10 @@
 import React, { useState } from 'react'
 import { useMutation } from 'react-query'
-import requestTaskApi from '../../config/axiosConfig'
+import requestTaskApi from '../../../config/axiosConfig'
 import { toast } from 'react-toastify'
 import { useNavigate } from 'react-router-dom'
 import { store } from '../../../redux/store'
-import { getClientResponse } from '../../../redux/clientResponse'
+import { getClientResponse } from '../../../redux/user'
 
 const useLoginForm = () => {
 	const navigate = useNavigate()
