+++ src/redux/user.ts
@@ -1,0 +1,22 @@
+import { createSlice, PayloadAction } from '@reduxjs/toolkit'
+
+export type StatusCode = {
+	isLoggedIn: boolean;
+};
+
+const initialState: StatusCode = { isLoggedIn: true }
+
+export const user = createSlice({
+	name: 'clientResponse',
+	initialState,
+	reducers: {
+		getClientResponse: (
+			state,
+			action: PayloadAction<{ isLogged: boolean }>
+		) => {
+			state.isLoggedIn = action.payload.isLogged
+		},
+	},
+})
+
+export const { getClientResponse } = user.actions
